requirements.txt
flask
langchain
langchain-mistralai
tavily-python
beautifulsoup4
python-dotenv
requests